consumer-system/adapter/requirements.txt
flask==2.3.3
flask-cors==4.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
httpx==0.25.0
pymysql==1.1.0
pika==1.3.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
cryptography==41.0.4
python-jose==3.3.0
requests==2.31.0
schedule==1.2.0
uvicorn
pymysql
python-keycloak